+++ app/src/main/java/pro/kisscat/www/bookmarkhelper/converter/support/pojo/App.java
@@ -19,7 +19,7 @@ public class App {
     public String name = "";
     public String packageName = "";
     public String versionName = "";
-    public int versionCode = 0;
+    public long versionCode = 0;
     @JSONField(serialize = false)
     public transient Drawable icon;
 }
